poc/requirements.txt
pykrx>=1.0.45
yfinance>=0.2.40
pandas>=2.0.0
sqlalchemy>=2.0.0
python-dateutil>=2.8.2
tqdm>=4.66.0
# ML 실험용
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
matplotlib>=3.7.0
tabulate>=0.9.0
# Postgres 마이그레이션
psycopg[binary]>=3.2.0
